re_query_tool/requirements.txt
django-tastypie==0.12.1
django==1.6
